Pass update check type to shouldProceedWithUpdate delegate method
