+++ Sparkle/SPUBasicUpdateDriver.h
@@ -8,6 +8,7 @@
 
 #import <Foundation/Foundation.h>
 #import "SPUUpdateDriver.h"
+#import "SPUUpdateCheck.h"
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -28,7 +29,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface SPUBasicUpdateDriver : NSObject
 
-- (instancetype)initWithHost:(SUHost *)host updater:(id)updater updaterDelegate:(nullable id <SPUUpdaterDelegate>)updaterDelegate delegate:(id <SPUBasicUpdateDriverDelegate>)delegate;
+- (instancetype)initWithHost:(SUHost *)host updateCheck:(SPUUpdateCheck)updateCheck updater:(id)updater updaterDelegate:(nullable id <SPUUpdaterDelegate>)updaterDelegate delegate:(id <SPUBasicUpdateDriverDelegate>)delegate;
 
 - (void)setCompletionHandler:(SPUUpdateDriverCompletion)completionBlock;
